msm7x30: Camera HAL for HTC Sense devices (1/2)
The proprietary camera files must be from an HTC Sense build for this
HAL to function properly. These files include:

libcamera.so
libmmipl.so
libmmjpeg.so
liboemcamera.so

Panorama and CAF camera application natively supported
Front facing camera and face unlock supported, if available

Touched up for jellybean.

Modified so it doesnt mess with other 7x30 targets not using this.

Change-Id: I12e73b0847ad80caec4b28157146047567c20df3